prevent from creating too many jobs by accident

+++ batch_purdue_pbs/resubmit_cache.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+MAX_JOBS=300
 queue=${1:-"cms"}
 perjob=${2:-20}
 
@@ -19,10 +20,15 @@ python chunk_submits.py $perjob "$SUBMIT_DIR" "$jobfiles_path" "re_jobfiles.txt"
 #Split on line, not on space
 IFS=$'\n'
 
+njobs=$(wc -l jobfiles_merged.txt | awk '{ print $1 }')
+if [ njobs \> $MAX_JOBS ]; then
+    echo "You are trying to create $njobs jobs, and the threshold is $MAX_JOBS. To override this, change the value of  MAX_JOBS in resubmit_cache.sh"
+    return
+fi
+
 for f in `cat re_jobfiles_merged.txt`; do
     rm cache_.sub
     cat template.sub > cache_.sub
     echo "./hmm_cache.sh $f" >> cache_.sub
     qsub cache_.sub -q $queue
-#    break # to test on 1 job
 done
