+++ batch_purdue_pbs/submit_analyze.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+MAX_JOBS=300
 queue=${1:-"cms"}
 perjob=${2:-20}
 
@@ -19,10 +20,16 @@ python chunk_submits.py $perjob "$SUBMIT_DIR" "$jobfiles_path" "jobfiles.txt"> j
 #Split on line, not on space
 IFS=$'\n'
 
+njobs=$(wc -l jobfiles_merged.txt | awk '{ print $1 }')
+if [ njobs \> $MAX_JOBS ]; then
+    echo "You are trying to create $njobs jobs, and the threshold is $MAX_JOBS. To override this, change the value of  MAX_JOBS in submit_analyse.sh"
+    return
+fi
+
 for f in `cat jobfiles_merged.txt`; do
     rm analyze_.sub
     cat template.sub > analyze_.sub
     echo "./hmm_analyze.sh $f" >> analyze_.sub
-    qsub analyze_.sub -q $queue
+#    qsub analyze_.sub -q $queue
 
 done
