+++ batch_purdue_pbs/submit_cache.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+MAX_JOBS=300
 queue=${1:-"cms"}
 perjob=${2:-20}
 
@@ -18,6 +19,12 @@ python chunk_submits.py $perjob "$SUBMIT_DIR" "$jobfiles_path" "jobfiles.txt" > 
 
 #Split on line, not on space
 IFS=$'\n'
+
+njobs=$(wc -l jobfiles_merged.txt | awk '{ print $1 }')
+if [ njobs \> $MAX_JOBS ]; then
+    echo "You are trying to create $njobs jobs, and the threshold is $MAX_JOBS. To override this, change the value of  MAX_JOBS in submit_cache.sh"
+    return
+fi
 
 for f in `cat jobfiles_merged.txt`; do
     rm cache_.sub
